refactor(components/atom/actionButton): api tab
